Manage the host interface with an FSM poked by events
Concentrating the host interface's main behaviour into an FSM feels
like the right thing to be doing.  Spreading it out over a large
number of functions will lead to difficult to debug code.  There are
three types of event that can occur currently:

 * EV_RX_FRAME_RECEIVED: Fired when a frame is received.
 * EV_RX_DONE: Fired when when the frame that was just received can be
	       discarded -- so we can write over the buffer it was in.
 * EV_TX_QUEUED: Fired when a frame has been placed in the output
 		 buffer for transmission.

A timeout event for ACKs from the host is to be added soon.

hostser.c now presents an API that can be used from the mainloop to
transmit and receive frames.  The next target is write the SRIC
interface in a similar manner and tie them together in the mainloop.