Fix data ablation: Add locking to the output buffer
The reception code could ablate a currently-under-construction
transmit buffer by transmitting an ACK to the host.  This patch
requires that the hostser FSM is moved into a 'locked' state in which
it is guaranteed that no interrupts will touch the transmit buffer.

This was the fastest route to achieving this.  The disadvantage of
this is that any frames that come in whilst the transmit buffer is
locked will be discarded (and later retransmitted by the host upon
timeout on the ACK).  The solution to this problem will be to have two
transmit buffers -- one for ACK'ing and one for general use.  The
ACK'ing buffer should be able to be const, which is nice...