Convert the USART1 driver into a generic USARTx driver
The code in usart.[hc] can drive both USART0 and USART1.  The
configuration information for each peripheral is provided to the
driver in a const struct containing callback addresses, and baud
rate.

This patch also modifies the hostser code to be configured to use a
given serial device.  This means that a driver with a similar
interface can be written for the USCI peripheral and plugged into the
hostser code without serial-device-specific code changes to the
hostser code.

Furthermore, the SRIC interface side of the PC-SRIC board will use
this generic USARTx driver soon :)