Pass command frames from hostser to sric bus
Adds function to the gateway state machine so that it will pass
commands that come in on hostser to the SRIC functions (which are
currently unimplemented), then transmit the SRIC response to the
host.

I've avoided adding any ACK'ing at this point.  This requires some
deep thought about what's going on...  I felt that the ACK'ing was
making it more complex than I wanted it to be right now.